refactor: Remove unused code, added namespaces Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
