--- typescript/vrotsc/tmp/compiler/transformers/scripts/namespaces.ts
@@ -1,136 +1,0 @@
-/*
- * #%L
- * vrotsc
- * %%
- * Copyright (C) 2023 VMware
- * %%
- * Build Tools for VMware Aria
- * Copyright 2023 VMware, Inc.
- * 
- * This product is licensed to you under the BSD-2 license (the "License"). You may not use this product except in compliance with the BSD-2 License.  
- * 
- * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
- * #L%
- */
-namespace vrotsc {
-    const ts: typeof import("typescript") = require("typescript");
-
-    export function collectNamespaces(sourceFile: ts.SourceFile, context: ScriptTransformationContext): ts.SourceFile {
-        const visitor = createVisitor(visitNode, context);
-
-        visitor.visitNode(sourceFile);
-
-        return sourceFile;
-
-        function visitNode(node: ts.Node): ts.VisitResult<ts.Node> {
-            switch (node.kind) {
-                case ts.SyntaxKind.ModuleDeclaration:
-                    visitModuleDeclaration(<ts.ModuleDeclaration>node);
-                    break;
-            }
-            return undefined;
-        }
-
-        function visitModuleDeclaration(node: ts.ModuleDeclaration): void {
-            if (!hasModifier(node.modifiers, ts.SyntaxKind.ExportKeyword)) {
-                const name = getPropertyName(node.name);
-                if (name) {
-                    if (context.globalIdentifiers.indexOf(name) < 0) {
-                        context.file.hierarchyFacts |= HierarchyFacts.ContainsGlobalNamespace;
-                        context.globalIdentifiers.push(name);
-                    }
-                }
-            }
-        }
-    }
-
-    export function transformNamespaces(sourceFile: ts.SourceFile, context: ScriptTransformationContext): ts.SourceFile {
-        if (context.globalIdentifiers.length) {
-            const statements = sourceFile.statements.map(node => {
-                switch (node.kind) {
-                    case ts.SyntaxKind.VariableStatement:
-                        return transformNamespaceVariable(node as ts.VariableStatement);
-                    case ts.SyntaxKind.ExpressionStatement:
-                        return transformExpressionStatement(node as ts.ExpressionStatement);
-                }
-                return node;
-            });
-
-            return ts.updateSourceFileNode(
-                sourceFile,
-                ts.setTextRange(
-                    ts.createNodeArray([
-                        ...statements
-                    ]),
-                    sourceFile.statements));
-        }
-
-        return sourceFile;
-
-        function transformNamespaceVariable(node: ts.VariableStatement): ts.VariableStatement {
-            if (node.declarationList && node.declarationList.declarations && node.declarationList.declarations.length === 1) {
-                const varNode = node.declarationList.declarations[0];
-                if (!varNode.initializer) {
-                    const varName = getIdentifierTextOrNull(varNode.name);
-                    if (varName && context.globalIdentifiers.indexOf(varName) > -1) {
-                        return ts.updateVariableStatement(
-                            node,
-                            node.modifiers,
-                            ts.updateVariableDeclarationList(
-                                node.declarationList,
-                                [
-                                    ts.updateVariableDeclaration(
-                                        varNode,
-                                        varNode.name,
-                                        varNode.type,
-                                        /* initializer */
-                                        ts.createBinary(
-                                            /* left */
-                                            ts.createPropertyAccess(
-                                                /* expression */
-                                                ts.createIdentifier(SCRIPT_VRO_GLOBAL),
-                                                /* name */
-                                                ts.createIdentifier(varName),
-                                            ),
-                                            /* operator */
-                                            ts.createToken(ts.SyntaxKind.BarBarToken),
-                                            /* right */
-                                            ts.createParen(
-                                                ts.createBinary(
-                                                    /* left */
-                                                    ts.createPropertyAccess(
-                                                        /* expression */
-                                                        ts.createIdentifier(SCRIPT_VRO_GLOBAL),
-                                                        /* name */
-                                                        ts.createIdentifier(varName),
-                                                    ),
-                                                    /* operator */
-                                                    ts.createToken(ts.SyntaxKind.FirstAssignment),
-                                                    /* right */
-                                                    ts.createObjectLiteral([], false),
-                                                )
-                                            ),
-                                        )),
-                                ]
-                            ));
-                    }
-                }
-            }
-            return node;
-        }
-
-        function transformExpressionStatement(node: ts.ExpressionStatement): ts.ExpressionStatement {
-            if (node.expression.kind === ts.SyntaxKind.CallExpression) {
-                const callNode = node.expression as ts.CallExpression;
-                if (callNode.arguments.length === 1 && callNode.arguments[0].kind === ts.SyntaxKind.BinaryExpression) {
-                    const firstCallArg = callNode.arguments[0] as ts.BinaryExpression;
-                    const ns = getIdentifierTextOrNull(firstCallArg.left);
-                    if (ns && context.globalIdentifiers.indexOf(ns) > -1) {
-                        (callNode.arguments[0] as any) = firstCallArg.left;
-                    }
-                }
-            }
-            return node;
-        }
-    }
-}
